requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.0.0
lightgbm>=3.3.0
catboost>=1.0.0

# Deep Learning (optional, for LSTM)
torch>=1.10.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# SHAP explanations
shap>=0.40.0

# Feature selection
boruta>=0.3  # Optional, for Boruta feature selection

# Statistics
statsmodels>=0.13.0

# Data handling
openpyxl>=3.0.0  # For Excel file reading
pyarrow>=6.0.0  # For Parquet files

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Utilities
tqdm>=4.62.0
joblib>=1.1.0
